Create Bloc, and Service Locator

+++ lib/movies/data/models/movies_model.dart
@@ -18,9 +18,7 @@ class MoviesModel extends Movies {
       backdropPath: json['backdrop_path'],
       genreIds: List<int>.from(json['genre_ids'].map((e) => e)),
       overview: json['overview'],
-
-      /// TODO : Check this
-      voteAverage: json['vote_average'],
+      voteAverage: json['vote_average'].toDouble(),
       releaseDate: json['release_date'],
     );
   }
